% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{cooks.dist}
\alias{cooks.dist}
\title{Cook's distance}
\usage{
function(model,tau=0, npoints=0, dist="t", plot=FALSE,xlab="Index",ylab="Cook distance",ylim=c(0,1),pch=19,cex=0.5,type="h",col="black")
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit.}

\item{npoints}{the maximum number of points to be identified.}

\item{dist}{assumed distribution for the dependent variable y.}

\item{plot}{logical: if plot is TRUE, the estimate weight against MT residual are plotted and if FALSE the weights are printed.}

\item{type}{1-character string giving the type of plot desired.}
}
\description{
A measure that combines the information of leverage and residual of the observation
}
\examples{
 
data("PSID1976")
y<-PSID1976$wage
x1<- PSID1976$age
x2<- PSID1976$education
x3<- PSID1976$youngkids
x4<- PSID1976$oldkids
x5<- PSID1976$experience

form <- (y ~ x1+x2+x3+x4+x5)
mt<- tobit(form, dist="t") #tobit-t model
mnormal <- tobit(form) # tobit normal model

#Analysis under the normal tobit model

cooks.dist(mnormal,npoints=1,dist="normal",ylim=c(0,0.6),plot=TRUE)

#Analysis under the tobit t model 

cooks.dist(mt,npoints=0,dist="t",ylim=c(0,0.6),plot=TRUE)

}
\references{
Barros, M., Galea, M., Leiva, V. and Santos-Neto, M. Generalized tobit models: Diagnostics and application in econometrics. (submitted manuscript)
}

