% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{cooks.distn}
\alias{cooks.distn}
\title{Cook's distance}
\usage{
cooks.distn(formula, tau = 0, npoints = 0, dist = "t", plot = FALSE,
  xlab = NULL, ylab = NULL, ylim = NULL, xlim = NULL, pch = 19,
  cex = 0.5, type = "p", col = NULL)
}
\arguments{
\item{tau}{x}

\item{npoints}{x}

\item{dist}{x}

\item{plot}{x}

\item{model}{x}
}
\description{
A measure that combines the information of leverage and residual of the observation
}

