% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{diag.tobit}
\alias{diag.tobit}
\title{Local Influence}
\usage{
function(model,tau=0,npoints=0,dist="t",perturbation=c("cases","scale","response","explanatory"),l=NULL,ylim=c(0,0.40),plot.ci=FALSE,plot.dmax=FALSE,vecplot = c("theta","beta","sigma"))
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit.}

\item{tau}{is the censoring point. The default is zero.}

\item{npoints}{pthe maximum number of points to be identified.}

\item{dist}{assumed distribution for the dependent variable y.}

\item{perturbation}{types of perturbation schemes for to be used.}

\item{l}{default is NULL. Necessary for the explanatory variable perturbation.}

\item{ylim}{the y limits of the plot.}

\item{plot.ci}{a logical value indicating if the normal curvature should be plotted or no.}

\item{plot.dmax}{a logical value indicating if dmax should be plotted or no.}

\item{vecplot}{informe if the normal curvature is calculed for theta, beta or sigma.}
}
\description{
Local influence plots for tobit model.
}
\examples{

data("PSID1976")
y<-PSID1976$wage
x1<- PSID1976$age
x2<- PSID1976$education
x3<- PSID1976$youngkids
x4<- PSID1976$oldkids
x5<- PSID1976$experience

form <- (y ~ x1+x2+x3+x4+x5)
mt<- tobit(form, dist="t") #tobit-t model
mnormal <- tobit(form) # tobit normal model

#Analysis under the normal tobit model

diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "cases",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under case-weight perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "cases",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under case-weight perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "cases",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under case-weight perturbation

diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "response",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under response perturbation
diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "response",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under response perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "response",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under response perturbation

diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "scale",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under scale perturbation
diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "scale",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under scale perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "scale",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under scale perturbation

diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="theta") #normal curvature for theta under explanatory perturbation
diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="beta") #normal curvature for beta under explanatory perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under explanatory perturbation

#Analysis under the tobit t model

diag.tobit(mt,npoints = 0,dist="t",perturbation = "cases",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under case-weight perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "cases",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under case-weight perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "cases",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under case-weight perturbation

diag.tobit(mt,npoints = 0,dist="t",perturbation = "response",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under response perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "response",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under response perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "response",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under response perturbation

diag.tobit(mt,npoints = 2,dist="t",perturbation = "scale",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under scale perturbation
diag.tobit(mt,npoints = 2,dist="t",perturbation = "scale",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under scale perturbation
diag.tobit(mt,npoints = 2,dist="t",perturbation = "scale",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under scale perturbation

diag.tobit(mt,npoints = 0,dist="t",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="theta") #normal curvature for theta under explanatory perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="beta") #normal curvature for beta under explanatory perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under explanatory perturbation

}
\references{
Barros, M., Galea, M., Leiva, V. and Santos-Neto, M. Generalized tobit models: Diagnostics and application in econometrics. (submitted manuscript)
}

