% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{diag.tobitn}
\alias{diag.tobitn}
\title{Local Influence Measures}
\usage{
diag.tobitn(model, tau = 0, npoints = 0)
}
\arguments{
\item{model}{x}

\item{tau}{x}

\item{npoints}{x}
}
\value{
x
}
\description{
x
}

