% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{Envelope}
\usage{
envelope(model, res = "martingalet", nboot = 19, alpha = 0.05, tau = 0,
  intercept = "TRUE", td = "t")
}
\arguments{
\item{model}{x}

\item{res}{x}

\item{nboot}{x}

\item{alpha}{x}

\item{intercept}{x}

\item{td}{x}
}
\value{

}
\description{
x
}

