% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{residuals}
\alias{residuals}
\title{Residuals}
\usage{
residuals(model, type, tau = 0, dist = "t")
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit.}

\item{type}{what type of residuals should be used. 
Possible type are: "deviance", "martingale" and "martingale type".}

\item{tau}{is the censoring point. The default is zero.}

\item{td}{assumed distribution for the dependent variable y. The default is student-t distribution.}
}
\value{
The returned object is a vector with one element for each subject.
}
\description{
Calculates martingale, deviance, martingale type residuals for tobit-t model.
}

