% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R, R/envelope.R
\name{tobitdiag}
\alias{tobitdiag}
\alias{diag.tobit}
\alias{cooks.dist}
\alias{AICc}
\alias{weights}
\alias{residuals}
\alias{envelope}
\title{Local Influence}
\usage{
diag.tobit(model,tau=0, npoints = 0,dist="t",
perturbation=c("cases","scale","response","explanatory"),
l=NULL,ylim=c(0,0.40),plot.ci=FALSE,plot.dmax=FALSE,
vecplot = c("theta","beta","sigma"))

cooks.dist(model, tau=0, npoints = 0, dist="t", 
plot=FALSE, xlab = "Index", ylab="Cook distance",
ylim=c(0,1), pch = 19, cex = 0.5, type="h", col="black")

AICc(object, object1 = NULL)

weights(model, plot = FALSE, npoints = 0)

residuals(model,type,tau=0,dist ="t")

envelope(model,res="martingalet",nboot = 19,alpha=0.05,tau=0,
intercept = "TRUE",dist="t",col="gray80",xlab="Q",ylab="R")
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit. The formula should be a symbolic description of a regression model of type \code{y ~ x1 + x2 + ... + xp}.}

\item{tau}{is the censoring point. The default is zero.}

\item{npoints}{the maximum number of points to be identified.}

\item{dist}{assumed distribution for the dependent variable y.}

\item{perturbation}{types of perturbation schemes for to be used.}

\item{l}{default is NULL. Necessary for the explanatory variable perturbation.}

\item{ylim}{the y limits of the plot.}

\item{plot.ci}{a logical value indicating if the normal curvature should be plotted or no.}

\item{plot.dmax}{a logical value indicating if dmax should be plotted or no.}

\item{vecplot}{informe if the normal curvature is calculed for theta, beta or sigma.}

\item{plot}{logical: if plot is TRUE, the estimate weight against MT residual are plotted and if FALSE the weights are printed.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{pch}{a vector of plotting characters or symbols: see \code{\link{points}}.}

\item{cex}{a numerical vector giving the amount by which plotting characters and symbols should be scaled relative to the default.}

\item{type}{what type of residuals should be used. 
Possible type are: "deviance", "martingale" and "martingale type".}

\item{col}{The colors for points. The default is gray80.}

\item{object}{a fitted model object for which there exists a logLik method to extract the corresponding log-likelihood, or an object inheriting from class logLik.}

\item{object1}{optionally more fitted model objects.}

\item{res}{character string indicating the type of residual 
desired. Possible values are "deviance", "martingale" and "martingale type".}

\item{nboot}{Number of simulated point patterns to be generated when computing the envelopes.}

\item{alpha}{the confidence level required. The default is to find 95 confidence envelopes.}

\item{intercept}{logical. Should an intercept be included in the null model?}
}
\value{
If just one object is provided, a numeric value with the corresponding AICc.

If two objects are provided, a data.frame with rows corresponding to the objects and columns 
representing the number of 
parameters in the model (df) and the AICc.

The returned object is a vector with one element for each subject.
}
\description{
Local influence plots for tobit model.

This function can be used to compute the Cook's distance.

AICc is AIC with a correction for finite sample sizes.

Estimate weight in the maximum likelihood estimation procedure under the tobit-t model.

Calculates martingale, deviance, martingale type residuals for normal tobit and tobit-t models.

Computes simulation envelopes of a normal tobit and tobit-t model.
}
\examples{

data("PSID1976")
y<-PSID1976$wage
x1<- PSID1976$age
x2<- PSID1976$education
x3<- PSID1976$youngkids
x4<- PSID1976$oldkids
x5<- PSID1976$experience

form <- (y ~ x1+x2+x3+x4+x5)
mt<- tobit(form, dist="t") #tobit-t model
mnormal <- tobit(form) # normal tobit model

#Analysis under the normal tobit model

diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "cases",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under case-weight perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "cases",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under case-weight perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "cases",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under case-weight perturbation

diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "response",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under response perturbation
diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "response",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under response perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "response",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under response perturbation

diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "scale",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under scale perturbation
diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "scale",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under scale perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "scale",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under scale perturbation

diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="theta") #normal curvature for theta under explanatory perturbation
diag.tobit(mnormal,npoints = 0,dist="normal",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="beta") #normal curvature for beta under explanatory perturbation
diag.tobit(mnormal,npoints = 3,dist="normal",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under explanatory perturbation

#Analysis under the tobit t model

diag.tobit(mt,npoints = 0,dist="t",perturbation = "cases",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under case-weight perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "cases",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under case-weight perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "cases",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under case-weight perturbation

diag.tobit(mt,npoints = 0,dist="t",perturbation = "response",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under response perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "response",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under response perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "response",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under response perturbation

diag.tobit(mt,npoints = 2,dist="t",perturbation = "scale",plot.ci=TRUE,vecplot="theta") #normal curvature for theta under scale perturbation
diag.tobit(mt,npoints = 2,dist="t",perturbation = "scale",plot.ci=TRUE,vecplot="beta") #normal curvature for beta under scale perturbation
diag.tobit(mt,npoints = 2,dist="t",perturbation = "scale",plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under scale perturbation

diag.tobit(mt,npoints = 0,dist="t",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="theta") #normal curvature for theta under explanatory perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="beta") #normal curvature for beta under explanatory perturbation
diag.tobit(mt,npoints = 0,dist="t",perturbation = "explanatory",l=2,plot.ci=TRUE,vecplot="sigma") #normal curvature for sigma under explanatory perturbation


 
data("PSID1976")
y<-PSID1976$wage
x1<- PSID1976$age
x2<- PSID1976$education
x3<- PSID1976$youngkids
x4<- PSID1976$oldkids
x5<- PSID1976$experience

form <- (y ~ x1+x2+x3+x4+x5)
mt<- tobit(form, dist="t") #tobit-t model
mnormal <- tobit(form) # normal tobit model

#Analysis under the normal tobit model

cooks.dist(mnormal,npoints=1,dist="normal",ylim=c(0,0.6),plot=TRUE)

#Analysis under the tobit t model 

cooks.dist(mt,npoints=0,dist="t",ylim=c(0,0.6),plot=TRUE)

 
data("PSID1976")
y<-PSID1976$wage
x1<- PSID1976$age
x2<- PSID1976$education
x3<- PSID1976$youngkids
x4<- PSID1976$oldkids
x5<- PSID1976$experience

form <- (y ~ x1+x2+x3+x4+x5)
mt<- tobit(form, dist="t") #tobit-t model
mnormal <- tobit(form) # normal tobit model

AICc(mnormal,mt)


 data("PSID1976")
 y<-PSID1976$wage
 x1<-PSID1976$age
 x2<- PSID1976$education
 x3<- PSID1976$youngkids
 x4<- PSID1976$oldkids
 x5<- PSID1976$experience 
 
 form <- (y ~ x1+x2+x3+x4+x5)
 #### tobit-t model ####
 mt<- tobit(form, dist="t") 
 
 weights(mt,plot=TRUE,npoints = 5)
 

data("PSID1976")
y<-PSID1976$wage
x1<- PSID1976$age
x2<- PSID1976$education
x3<- PSID1976$youngkids
x4<- PSID1976$oldkids
x5<- PSID1976$experience

form <- (y ~ x1+x2+x3+x4+x5)
mt<- tobit(form, dist="t") #tobit-t model
mnormal <- tobit(form) # normal tobit model

#Analysis under the normal tobit model
rest <- residuals(mnormal,"martingalet",dist="normal")
plot(rest,pch=16,cex=0.7, ylim=c(-3,6), ylab="R",xlab="I")
abline(h=c(-3,3),lty=2)
identify(rest,n=7,cex=1.0)

#Analysis under the tobit t model
rest <- residuals(mt,"martingalet",dist="t")
plot(rest,pch=16,cex=0.7, ylim=c(-3,6), ylab="R",xlab="I")
abline(h=c(-3,3),lty=2)


data("PSID1976")
y<-PSID1976$wage
x1<- PSID1976$age
x2<- PSID1976$education
x3<- PSID1976$youngkids
x4<- PSID1976$oldkids
x5<- PSID1976$experience

form <- (y ~ x1+x2+x3+x4+x5)
mt<- tobit(form, dist="t") #tobit-t model
mnormal <- tobit(form) # tobit normal model

#Analysis under the normal tobit model
envelope(mnormal,dist = "normal")

#Analysis under the tobit-t model
envelope(mt,dist = "t")
}
\references{
Barros, M., Galea, M., Leiva, V. and Santos-Neto, M. Generalized tobit models: Diagnostics and application in econometrics. (submitted manuscript)

Barros, M., Galea, M., Leiva, V. and Santos-Neto, M. Generalized tobit models: Diagnostics and application in econometrics. (submitted manuscript)

Barros, M., Galea, M., Leiva, V. and Santos-Neto, M. Generalized tobit models: Diagnostics and application in econometrics. (submitted manuscript)

Barros, M., Galea, M., Leiva, V. and Santos-Neto, M. Generalized tobit models: Diagnostics and application in econometrics. (submitted manuscript)

Barros, M., Galea, M., Leiva, V. and Santos-Neto, M. Generalized tobit models: Diagnostics and application in econometrics. (submitted manuscript)
}
