% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{weights}
\alias{weights}
\title{weights}
\usage{
weights(model, plot = FALSE, npoints = 0)
}
\arguments{
\item{model}{x}

\item{plot}{x}

\item{npoints}{x}

\item{tau}{x}

\item{dist}{x}
}
\description{
A measure that combines the information of leverage and residual of the observation
}

