% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{weights.tobit}
\alias{weights.tobit}
\title{Weights}
\usage{
\method{weights}{tobit}(model, plot = FALSE, res = "martingalet", tau = 0,
  ...)
}
\arguments{
\item{model}{x}

\item{plot}{x}

\item{tau}{x}

\item{npoints}{x}

\item{dist}{x}
}
\description{
A measure that combines the information of leverage and residual of the observation
}

