% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{diag.tobit}
\alias{diag.tobit}
\title{Local Influence}
\usage{
diag.tobit(model, tau = 0, npoints = 0, dist = "t",
  perturbation = c("cases", "scale", "response", "explanatory"), l = NULL,
  ylim = c(0, 0.4), plot.ci = FALSE, plot.dmax = FALSE,
  vecplot = c("theta", "beta", "sigma"))
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit.}

\item{tau}{is the censoring point. The default is zero.}

\item{npoints}{points to be plotted.}
}
\description{
Local influence plots for tobit model.
}

