% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{diag.tobitn}
\alias{diag.tobitn}
\title{Local Influence}
\usage{
diag.tobitn(model, tau = 0, npoints = 0)
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit.}

\item{tau}{is the censoring point. The default is zero.}

\item{npoints}{points to be plotted.}
}
\description{
Local influence plots for tobit model.
}

