% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtobit.R
\name{diag.tobitt}
\alias{diag.tobitt}
\title{Local Influence}
\usage{
diag.tobitt(model, tau = 0, npoints = 0, perturbation = c("cases",
  "scale", "response", "explanatory"), l = NULL, ylim = c(0, 0.4),
  plot.ci = FALSE, plot.dmax = FALSE, vecplot = c("theta", "beta",
  "sigma"))
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit.}

\item{tau}{is the censoring point. The default is zero.}

\item{npoints}{points to be plotted.}
}
\description{
Local influence plots for tobit-t model.
}

