% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope.normal}
\alias{envelope.normal}
\title{Envelope}
\usage{
envelope.normal(model, nboot = 19, alpha = 0.05, tau = 0,
  intercept = "TRUE")
}
\arguments{
\item{model}{x}

\item{nboot}{x}

\item{alpha}{x}

\item{tau}{x}

\item{intercept}{x}
}
\value{

}
\description{
x
}

