% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{Envelope}
\usage{
envelope(model, res = "martingalet", nboot = 19, alpha = 0.05, tau = 0,
  intercept = "TRUE", td = "t")
}
\arguments{
\item{model}{an object of class "tobit" as fitted by tobit.}

\item{res}{character string indicating the type of residual 
desired. Possible values are "deviance", "martingale" and "martingale type".}

\item{nboot}{Number of simulated point patterns to be generated when computing the envelopes.}

\item{alpha}{the confidence level required. The default is to find 95 confidence envelopes.}

\item{intercept}{logical. Should an intercept be included in the null model?}

\item{td}{assumed distribution for the dependent variable y.}
}
\description{
Computes simulation envelopes of a tobit-t model.
}

